doc_flow/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
boto3==1.36.8
botocore==1.36.8
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
distro==1.9.0
elevenlabs==1.50.5
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.6
git-filter-repo==2.47.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jmespath==1.0.1
lxml==5.3.0
openai==1.57.4
openpyxl==3.1.5
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.8.0
pydantic_core==2.27.1
PyMuPDF==1.25.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
python-pptx==1.0.2
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.34.0
websockets==14.2
XlsxWriter==3.2.0
